pyproject.toml
[project]
name = "ipad-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers>=0.36.0",
    "einops>=0.8.1",
    "lightning>=2.6.0",
    "mlflow>=3.8.0",
    "omegaconf>=2.3.0",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "safetensors>=0.7.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[dependency-groups]
dev = [
    "accelerate>=1.12.0",
    "dvc[webdav]>=3.65.0",
    "fire>=0.7.1",
    "img2dataset>=1.47.0",
    "ipykernel>=7.1.0",
    "pre-commit>=4.5.1",
]
